Move reg-only jax tests from test_objective_jax to test_regularizers_jax
